Time: 86 ms (25.84%) | Memory: 86.6 MB (20.97%) - LeetSync

+++ 2113-find-the-kth-largest-integer-in-the-array/find-the-kth-largest-integer-in-the-array.cpp
@@ -2,9 +2,10 @@ class Solution {
 public:
     string kthLargestNumber(vector<string>& nums, int k) {
         priority_queue<pair<int,string>,
-        vector<pair<int,string>>,greater<pair<int,string>>>pq;
+        vector<pair<int,string>>,
+        greater<pair<int,string>>>pq;
         for(string s:nums){
-            pq.push({(int)s.length(),s});
+            pq.push({s.length(),s});
             if(pq.size()>k){
                 pq.pop();
             }
@@ -12,4 +13,4 @@ public:
         return pq.top().second;
     }
 };
-        
+
